made some changes in and out of the 04-flex folder
So made some changes in style.css and added/commited when being first in
the 04-flex folder, then went out of the folder, in the CSS-stuff, and
cause vscode had that style.css still opne, did more changes. Lets see
if ALL changes are kept if I add/commit and push ONLY from CSS-stuff
folder. As when added and commited in the CSS folder, then went in the
04-flex folder, and git status showed the style.css in red as modified,
while in the CSS-style folder git status showed the path 04-flex/style.
css as green and added then commited,