Changes to file from dir above
So went out of the directory that contained this html file
and changed it as it was still open in vscode (just added a commnet)
Then in cmd line cd .. out od the 04-flex-tkse-2, and commited it
with this message. Lets see, if I push it from this (CSS-stuff) dir
if I will be able later to push from inside the 0-flex-take directory.